First Commit
These notebooks are used to calculate the effects of DC electric fields
on the hydrogen atom. To do this, the hydrogen wavefunctions, matrix
elements and energies are required. The energies are calculated by
Horbatsch and Hessels. The matrix elements are calculated analytically
for |n,l,m> states in MatrixElements.nb. These elements are then
converted to hyperfine matrix elements by the Wigner-Eckhart theorem and
used to calculate the effect of a perturbing electric field.

The main notebook DCStarkv2.nb can also be used to calculate the
hydrogen state lifetimes.